Implemented GPU_RAM_Layout class
The idea is to put 2 blocks of degree of freedom trajectories on the GPU
and calculate the entropy for every pair of degrees of freedom. Space is
needed to store these two blocks, the calculation results, the histograms
and their numbers of occupied bins. The maximum number of degrees of freedom
which fits in the GPU RAM is calculated and the RAM is partitioned
accordingly using pointers.