Reworked CPU_RAM_Block and GPU_RAM_Block classes
The idea of the CPU_RAM_Block class is to manage a chunk
of degrees of freedom. Upon creation, the object is set up
virtually at first. Only when calling the deploy method,
the according data is actually loaded from the harddisk.
At this time, GPU_RAM_Blocks are created, also only virtually
at first. Upon calling the GPU_RAM_Blocks deploy method,
the data is transferred to the GPU RAM.