Implemented consistent naming scheme for dof vars
a _g suffix means that the variable is global. If there are 1000 atoms,
there are 3 * 1000 - 6 = 2994 dofs(degrees of freedom). Thus,
a _g variable would range from 0 to 2993. A _l suffix means that the
variable is local to the block. E.g., if the block starts at dof 300
and dof_g = 400, then dof_l = 100. A _t suffix (type) means that the
variable either refers to only bonds, only angles or only dihedrals.
In our example, there are 1000 - 1 = 999 bonds and 1000 - 2 = 998 angles.
So, if dof_g = 1050, then dof_gt = 1050 - 999 = 51.
The order is bonds before angles, angles before dihedrals.