Added fast-math switch to PARENT_GPU_kernels.cu
By uncommenting the line "// #define FAST_MATH",
the gpu kernels are compiled to use the CUDA __logf
and __sinf functions. In double precision, neither
does this switch significantly improve performance,
nor does it significantly alter the result.
From the performance side, this should mean that the
CUDA Kernels are memory bound. I will investigate
this closer when I run a profiler on the code during
the code optimization phase.