Implemented mutual information masking
When calculatting the maximum information
spanning tree approximation using Prim's
algorithm, one needs to keep track of
which degrees of freedom have already
been added to the tree. This is done
by decomposing the index of the mutual
information term into the indices of the
two degrees of freedom which share this
mutual information. If only one of the two
dofs has already been added to the tree
(determined by their dof_flags), then
the mutual information is processed
normally. Otherwise it is set to zero
(mutual information is non-negative
definite) and thus taken out of the
calculation.

modified:   src/MIST_GPU/MIST_GPU.cu