Implemented openMP baseline for MIST
In order to achieve better performance,
the data preprocessing is shifted to the
CPU. For this, a pure openMP version was
implemented as a baseline. Depending on
the size of the array, the reduction will
be performed either on the CPU or the GPU
in the future version of the code.

modified:   compile_run_MIST.sh
modified:   src/MIST_GPU/MIST_GPU.cu
copied:     src/MIST_GPU/MIST_GPU.cu -> src/MIST_GPU/MIST_GPU.cu.sav
new file:   src/MIST_GPU/MIST_openMP.cpp