First high performance max info implementation
Compared with my MPI/openMP version, the GPU
maximum inforamtion spanning tree approximation
code on a GTX 2060 super is as fast as 83!
Intel(R) Xeon(R) CPU E5-2660 0 @ 2.20GHz physical
cores. This estimation is based on the MPI/openMP
calculation times of 4 of these processors
(8 cores each) which are connected via InfiniBand.
The GPU version calculates in less than 40% of the
time.