Optimized histograms in PARENT_GPU
Boundary checks in the histogram kernel were
taken out. Shared memory use was implemented.
However, when binning to 2500 bins, which
is common for entropy calculations, the
shared memory version is hardly more performant
than writing directly to global memory via
atomicAdd. The reason is that write collisions
do not occur frequently with 2500 bins.