Implemented gbat trajectory format
src/util/classes/Bat.cpp was enhanced to convert
from .bat to .gbat format. With the new .gbat format,
degrees of freedom are stored compact, i. e.
the frames of a degree of freedom are stored next
to each other. This will enhance harddisk reading
speed for PARENT_GPU.cu and should significantly
improve performance if the whole trajectory
does not fit into CPU RAM and reloading is necessary.