Implemented RAM switches in PARENT_GPU.cu
The user can now specify how much CPU
as well as GPU RAM he wants to provide
at maximum for the calculation. For the
CPU RAM, if the uswer provides more than
what is needed, only the necessary amount
is allocated. The same cap will be
implemented for the GPU in the next commit.