Modified PARENT_GPU.cu to use less CPU RAM
Previously, the CPU_RAM_Layout class allocated
memory to store results for 2 gpu_block and
1 gpu_block_pair calculations simultaneously.
However, the results are copied to CPU ram after
each gpu_block/gpu_block_pair calculation.
As gpu_block_pair uses most ram, gpu_block
calculations fit into the ram for the gpu_block_pair
calculations, so only the ram for the gpu_block_pair
needs to be allocated. Furthermore, previously
space for reading the trajectory was allocated.
This is now accomplished by the Bat class, so
this memory does not need to be allocated separately
anymore.