Removed unused kernels from PARENT_GPU_kernels.cu
Kernels for 1D entropy calculations are not used,
as their workload as negligible compared to the 2D
entropy calculations. Therefore, those calculations
are done on the cpu, at least for now. Furthermore,
there were development kernels for calculating the
p lnp sum. 98% of the GPU kernel time is spent on
the histogram calculations. Therefore, optimizing
the p lnp sum kernels has low priority for now.