Implemented vectorized reads/writes in kernel
The PARENT_GPU_kernels.cu histogram kernel
now reads and writes larger chunks of memory.
To do so, padded trajectory reading was
implemented in Bat.cpp . The performance gain
is questionable due to the additional overhead.
Benchmarks are running. Depending on the results,
this branch might or might not be merged into the
master branch.