Bugfixed errors around convert_BAT_to_GBAT
In the Makefile, convert_BAT_to_GBAT.cpp is now
compiled with --std=c++11. Newer g++ versions
support gnu++14 (i. e. a superset of c++14) by
default, so this can also be considered a
legacy compatibility fix. The function
which was not recognized is to_string.
There another bug was fixed where this
function was used to produce an error
message and used integer instead of
floating point division.