Fixed bug in PARENT_GPU.cu for extremely low RAM
If the CPU RAM provided is extremely low compared
to the provided GPU RAM, there can be a single
GPU_RAM_Block in a CPU_RAM_Block. Then, a loop
is never executed which deploys this GPU_RAM_Block.
This is fixed by deploying the block outside
of the loop in this special case.

Also, shared memory histograms are turned off
for now, as they are prone to another bug
associated to moderateley low CPU RAM.