Fixed bug in PARENT_GPU.cu for odd GPU dofs
When shared memory histograms were used and
the degrees of freedom per block on the GPU
was odd, the CUDA kernels failed due to
misalignment. This is due to the different
variable sizes (4 bytes, 8 bytes ...).
This was fixed by changing the order of
the alloctated arrays of the GPU in
GPU_RAM_Layout. This also resolved
illegal memory accesses when the
dofs per block on the GPU were larger
than on the cpu.

Shared memory histograms are now enabled again.