Implemented Work class
Essentially, this class gets filled with GPU_RAM_Blocks
and pairs of GPU_RAM_Blocks to be calculated. Afterwards,
for each GPU a thread is spawned which obtains work
from this Class in a threadsafe manner.

The code works for a single GPU, but ist still slow.
Additionally, it needs to be debugged to work with
multiple GPUs.