Multi-GPU computation working
Multi-GPU computation has been implemented and
tested to be working. Speed increase is almost linear
with the cumulative GPU computational power.

The data fetching could still be optimized, the expected
speedup is rather minor however (based on the linear behavior
in the benchmarks and comparison with the single-GPU code).

While the results on my RTX 2060 Super are consistent, there
is some randomness in the results of my GTX 1060. This is
most probably due to the high parallelism and the
non-associativity of the binary floating point addition.
As a check, the plnp sum could be calculated on the CPU,
probably with only minor speed decrease, as the by far largest
part of the calculation time is spent during building the histograms.
Neverthless, also on the GTX 1060, the results are consistent
up to the second digit after the comma for the 1UGH complex
(3121 atoms), which is by far sufficient in precision. This
randomness on the GTX 1060 is also observed for the single-GPU code.