Added MIST_GPU_large.cu to deprecated folder
MIST_GPU.cu works for molecules where the mutual
information matrix fits entirely on the GPU.
MIST_GPU_large.cu chunks the MI matrix in this case
and therefore works for molecules of any given size.
However, by chunking, it gets terribly slow for
large molecules. The reason lies in the fact that
the whole matrix needs to be processed during
each iteration step in Prim's algorithm, and there
are as many iterations as degrees of freedom.
For example, a 30000 atom molecule features an MI
matrix of roughly 30 GB and 90000 degrees of freedom,
which amounts to 2.7 Peta(!)byte data transfer to
the GPU.

For such large molecules, the user is encouraged
to use MIST_openMP on a machine with a high count
of cores and to have patience. The other option,
if a CPU cluster is available, is to perform the
calculation using the MPI version of MIST in my
other repository "PARENT", which is targeted at
CPU clusters.