Updated README.md with picture of clustering
This picture demonstrates results of the clustering
of interaction networks in a Ubiquitin-UBM2 complex.
Ubiquitin is the goto model protein of NMR spectroscopy,
one of the two major tequniques of structural biology.
Therefore this picture should be very interesting
to the community. In fact, an upcoming publication
will include excitingly striking results for other
biomolecules.