Added src/analysis/analyze_residue.cpp
This program lists all atoms, bonds, angles and dihedrals
in a residue. It gives the 1D entropy values as well as
all mutual information terms separately. Additionally,
the cumulative 1D entropy and mutual information values
are listed per degree of freedom (pair) type. Furthermore,
the degree of freedom with the largest 1D entropy value
and the degree of freedom pair with the largest mutual
information value are listed.