Added src/analysis/analyze_residue_pair.cpp
This program does basically the same as
analyze_residue.cpp in the last commit,
but for pairs of resdiues. It lists all
the 1D details for both residues speparately,
but different to analyze_residue.cpp, it
lists the details of the mutual information
between the residue pairs, not the mutual information
within the residues. This way, the interaction
between resdidue pairs is analyzed.