Changed include_full switch Residue_Representation
The include_full switch in the Residue_Representation constructor
is responsible for the association of bonds, angles and dihedrals
(==dofs) with the residues. E.g. for a bond and include_full==true,
the bond is only associated with a residue if both atoms are
part of the residue. Previously, include_full==false meant
that dofs were attributed to a residue if at least half of the
atoms were part of the residue. Now, include_false adds a dof
to a residue if the leading atom (i. e. atom 2 for a bond,
atom 3 for an angle and atom 4 for a dihedral) is part of the
residue. The reasoning behind this stems from the building process
of the BAT coordinate system. There, one iteratively adds atoms to the
graph, and every atom comes with one bond, one angle and one dihedral.
The added atom in every step is the leading atom of the bond, the
angle and the dihedral, which is why these three dofs can be
associated with this atom.

This affects the programs analyze_residue.cpp, analyze_residue_pair.cpp
as well as hierarchical_residue_clusters.cpp. For the first two
programs, with include_full==false, now all dofs belong to exactly
one residue, and this is always set now. For hierarchical_residue_clusters.cpp,
a command line switch was added to control include_full. With
include_full==false, many backbone dofs are added, so the analysis
is sensitive to the backbone as well as the global orientation of the
residues with respect to their neighbors. include_full==true, on the
other hand, focuses on the side-chains.