fix: integer overflow for large vram gpus.
A bug was fixed which for more than 1310 dofs
per block on the GPU caused an integer overflow.
In particular, the bug affected the pointer
calulation for the histograms. For 50 bins,
the offset of the current histogram is
50**2 * 1310**2 = 4,29025e9, which is just
lower than 2**32 for unsigned int.
50**2 * 1311**2 = 4,2968025e9, on the other
hand, is juast larger than 2**32, causing the
overflow. Modern GPUs with their large VRAM
(e.g. 24G) are well capable of reaching these
number of dofs per gpu block.